requirements.txt
# Broadway Show Research System - Dependencies
# Last updated to avoid version conflicts

langchain-openai>=0.1.0
tavily-python>=0.3.0
python-dotenv>=1.0.0
openai>=1.0.0
streamlit>=1.29.0